ajustes na exibicao dos valores do dashboard

+++ projeto-individual/database/migrations/2022_06_29_110929_create_releases_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('release_type');
             $table->string('person');
             $table->string('description');
-            $table->decimal('amount');
+            $table->double('amount', 16,2);
             $table->date('due_date');
             $table->timestamps();
         });
